Correctifs logs
Former-commit-id: 2fd710ea078ce3f8c5f35dd7005cd1b165925342

+++ htdocs/bimpcommercial/objects/Bimp_Propal.class.php
@@ -954,7 +954,7 @@ class Bimp_Propal extends BimpComm
 
         $date_diff = 0;
         if (isset($new_data['date_livraison']) && $new_data['date_livraison'] !== $this->getData('date_livraison')) {
-            $date_diff = (int) BimpTools::getDateForDolDate($new_data['date_livraison']) - (BimpTools::getDateForDolDate($this->getData('date_livraison')));
+            $date_diff = (int) BimpTools::getDateForDolDate($new_data['date_livraison']) - (int) BimpTools::getDateForDolDate($this->getData('date_livraison'));
         }
 
         $now = date('Y-m-d H:i:s');
